+++ pandas/tests/indexes/test_base.py
@@ -2480,6 +2480,15 @@ class TestMixedIntIndex(Base):
         expected = Index([], dtype=object)
         tm.assert_index_equal(result, expected)
 
+    def test_index_repr_bool_nan(self):
+        # Tests for the Index representation,
+        # specifically for the case that includes bools and NANs
+        # GH32146
+        arr = Index([True, False, np.nan], dtype=object)
+        exp = arr.format()
+        out = ["True", "False", "NaN"]
+        assert out == exp
+
 
 class TestIndexUtils:
     @pytest.mark.parametrize(
